Former-commit-id: 3d23d0b26a2d903ccf5f2b38fba14dbc0215242b
Former-commit-id: 682c4c492904ed3d4ddf8210868ff30fb275ebd7

+++ kernel/kernel-masterdata-service/src/main/java/io/mosip/kernel/masterdata/entity/Location.java
@@ -45,7 +45,7 @@ public class Location extends BaseEntity implements Serializable {
 	private String name;
 
 	@Column(name = "hierarchy_level", nullable = false)
-	private int hierarchyLevel;
+	private short hierarchyLevel;
 
 	@Column(name = "hierarchy_level_name", nullable = false, length = 64)
 	private String hierarchyName;
